model/calculate_question: Add question including calculating
Add a question type in which the user has to recognize two numbers, calculate a
new one by applying a mathematical operation to those two numbers and translate
that new number into turkish. This is harder than both the recognize and
translate question types and trains both recognition and active recall at once.
It is also mentally more engaging due to the calculating aspect, which prevents
the user from just learning the answers by heart.

Signed-off-by: Jim Gerth <jim.frederic.gerth@gmail.com>