extensions/written_out: Handle exception in _blocksToWord better
Instead of having an optional return type for _blocksToWord because of the error
that can occur, when the number is too big, rather handle that error with the
new custom NumberTooBigException to make the return type non-optional. This
improves usability, as that error really should not happen in normal use, and if
it does, more information about it is supplied now. Also add an assert for the
same issue.

Signed-off-by: Jim Gerth <jim.frederic.gerth@gmail.com>