widgets/exercise_topic_body_base: Add base for exercise topic bodies
Add the abstract ExerciseTopicBodyBase (and -State) with common
functionality for widgets to be used as the body for the ExerciseTopic.
This was extracted from the _ExerciseTopicBody (and -State) in
widgets/topics/exercise_topic.dart, keeping the current question
management and answer checking logic, but leaving the actualy interface
to be built abstract.

A few improvements are included here as well, like parameterizing the
snackbar's height (and increasing it a bit), making the state variables
only readable (and not writable) or completely hiding them for
implementing classes, closing any potential snackbar still open in
deactivate rather than in dispose, which semantically makes a bit more
sense, and adding and rewriting documentation.

Signed-off-by: Jim Gerth <jim.frederic.gerth@gmail.com>